removed warnings

+++ Solution/Game/LevelFactory.cpp
@@ -500,7 +500,7 @@ void LevelFactory::LoadDefendables(XMLReader& aReader, tinyxml2::XMLElement* aLe
 		std::string propType;
 		aReader.ForceReadAttribute(entityElement, "propType", propType);
 		myCurrentLevel->myEntityFactory->CopyEntity(newEntity, propType);
-		int difficultHealth = newEntity->GetComponent<HealthComponent>()->GetHealth() * (2 - aDifficultScale);
+		int difficultHealth = int(newEntity->GetComponent<HealthComponent>()->GetHealth() * (2 - aDifficultScale));
 		newEntity->GetComponent<HealthComponent>()->Init(difficultHealth);
 
 		std::string defendName;
